Replace ingredient quantity from int to double

+++ ProjectDesignPattern4AL2/Classes/Sandwich.cs
@@ -3,10 +3,10 @@ namespace ProjectDesignPattern4AL2.Classes;
 public class Sandwich
 {
     public string Name { get; set; }
-    public Dictionary<Ingredient, int> Ingredients { get; set; }
+    public Dictionary<Ingredient, double> Ingredients { get; set; }
     public double Price { get; set; }
 
-    public Sandwich(string name, Dictionary<Ingredient, int> ingredients, double price)
+    public Sandwich(string name, Dictionary<Ingredient, double> ingredients, double price)
     {
         Name = name;
         Ingredients = ingredients;
